requirements.txt
attrs==21.4.0
Automat==20.2.0
cffi==1.15.0
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.7.1
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pymongo==4.0.1
pyOpenSSL==22.0.0
python-dotenv==0.19.2
queuelib==1.6.2
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
Twisted==22.1.0
typing-extensions==4.0.1
w3lib==1.22.0
zope.interface==5.4.0
